requirements.txt
graphrag==0.3.2
pandas==2.2.2
neo4j-rust-ext==5.24.0.0
PyYAML~=6.0.2
requests~=2.32.3
beautifulsoup4~=4.12.3
tiktoken~=0.7.0
neo4j~=5.24.0
python-dotenv~=1.0.1

